requirements.txt
flask
mysql-connector-python
python-dotenv
requests
google-generativeai
google-genai
numpy
